requirements.txt
fastapi==0.92.0
uvicorn==0.20.0
structlog==22.3.0
requests==2.28.2
python-dotenv==0.21.1
pytest-mock==3.10.0
pytest==7.2.2
pytest-cov==4.0.0
sentry-sdk==1.17.0
strawberry-graphql[fastapi]==0.176.0
